refactor: move function "kill_internal" to "db_util"
- move function "kill_internal" from "MongoInstance" to "db_util"
- rename function "kill_internal" to "killProcess"

As suggested by https://github.com/nodkz/mongodb-memory-server/pull/389#discussion_r496997760
